+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Fixes
 
 - Start the document server after client hooks run so the document root can be set with other Maze Runner configuration [484](https://github.com/bugsnag/maze-runner/pull/484)
+- Add missing Busnag-Span-Sampling header to Access-Control-Allow-Headers [485](https://github.com/bugsnag/maze-runner/pull/485)
 
 # 7.20.0 - 2023/02/28